disruption data
