+++ server/data/disruption/data/disruption-data-base.ts
@@ -1,0 +1,2 @@
+/** Stores the data inherent to this particular type of disruption. */
+export abstract class DisruptionDataBase {}
