+++ server/data/disruption/period/ends/ends-exactly.ts
