+++ server/data/disruption/period/ends/ends-when-alert-ends.ts
